Code simplified.
Updates to Rigid Model do not quite happen in real time, probably due to
cost of traversing the graph.

TODO: change endpoint to update every 15 min
TODO: Click on a graph node to enhance it and highlight the table row.
TODO: Generate overall metrics to characterize delay impact on the
network.